UI bugs fixed

+++ app/src/main/java/org/mcmikecamara/stormy/weather/Hour.java
@@ -34,7 +34,7 @@ public class Hour implements Parcelable {
         mSummary = summary;
     }
 
-    public double getTemperature() {
+    public int getTemperature() {
 
         return ((((int)Math.round(mTemperature))- 32)*5)/9;
     }
